refactor: split skills modules under line limits
