+++ apps/client/src/components/knowledge-base/components/SkillsTab.tsx
@@ -1,6 +1,6 @@
 import './SkillsTab.scss'
 
-import { App, Form } from 'antd'
+import { App } from 'antd'
 import React from 'react'
 import type { ReactNode } from 'react'
 import { useTranslation } from 'react-i18next'
@@ -9,22 +9,15 @@ import useSWR from 'swr'
 
 import type { ConfigResponse } from '@vibe-forge/types'
 
-import type { SkillHubItem, SkillSummary } from '#~/api.js'
-import {
-  getApiErrorMessage,
-  getConfig,
-  importSkillArchive,
-  installSkillHubItem,
-  installSkillsCliItem,
-  listSkills,
-  searchSkillsCli
-} from '#~/api.js'
+import type { SkillSummary } from '#~/api.js'
+import { getConfig, listSkills } from '#~/api.js'
+import { useSkillsCliModalController } from './@hooks/use-skills-cli-modal-controller'
+import { useSkillsTabActions } from './@hooks/use-skills-tab-actions'
 import { ProjectSkillsList } from './ProjectSkillsList'
 import { SkillArchiveInput } from './SkillArchiveInput'
 import { SkillMarketView } from './SkillMarketView'
 import { SkillRegistryModal } from './SkillRegistryModal'
 import { SkillsCliModal } from './SkillsCliModal'
-import type { SkillsCliFormValues } from './SkillsCliModal'
 import { SkillsTabActions } from './SkillsTabActions'
 import { TabContent } from './TabContent'
 import { ALL_REGISTRIES, filterProjectSkills } from './skill-hub-utils'
@@ -32,15 +25,6 @@ import type { SkillHubInstallFilter, SkillHubSortKey } from './skill-hub-utils'
 import { useSkillMarketFilters } from './use-skill-market-filters'
 import { useSkillMarketSearch } from './use-skill-market-search'
 import { useSkillRegistryModal } from './use-skill-registry-modal'
-
-const SKILLS_CLI_INITIAL_LIMIT = 100
-const SKILLS_CLI_LIMIT_STEP = 100
-const SKILLS_CLI_MAX_LIMIT = 500
-
-const trimOptionalString = (value: string | undefined) => {
-  const normalizedValue = value?.trim()
-  return normalizedValue == null || normalizedValue === '' ? undefined : normalizedValue
-}
 
 interface SkillsTabProps {
   leading?: ReactNode
@@ -82,20 +66,6 @@ export function SkillsTab({
   const { t } = useTranslation()
   const { message } = App.useApp()
   const navigate = useNavigate()
-  const importInputRef = React.useRef<HTMLInputElement | null>(null)
-  const [installingId, setInstallingId] = React.useState<string | null>(null)
-  const [importing, setImporting] = React.useState(false)
-  const [skillsCliOpen, setSkillsCliOpen] = React.useState(false)
-  const [skillsCliSearching, setSkillsCliSearching] = React.useState(false)
-  const [skillsCliLoadingMore, setSkillsCliLoadingMore] = React.useState(false)
-  const [skillsCliInstallingId, setSkillsCliInstallingId] = React.useState<string | null>(null)
-  const [skillsCliItems, setSkillsCliItems] = React.useState<SkillHubItem[]>([])
-  const [skillsCliError, setSkillsCliError] = React.useState<string | null>(null)
-  const [skillsCliHasMore, setSkillsCliHasMore] = React.useState(false)
-  const [skillsCliHasSearched, setSkillsCliHasSearched] = React.useState(false)
-  const [skillsCliLimit, setSkillsCliLimit] = React.useState(SKILLS_CLI_INITIAL_LIMIT)
-  const [skillsCliResetKey, setSkillsCliResetKey] = React.useState('')
-  const [skillsCliForm] = Form.useForm<SkillsCliFormValues>()
   const {
     data: skillsRes,
     isLoading: isSkillsLoading,
@@ -108,6 +78,19 @@ export function SkillsTab({
     mutateConfig,
     mutateHub: marketSearch.mutate,
     setRegistry: onRegistryChange
+  })
+  const actions = useSkillsTabActions({
+    marketMutate: marketSearch.mutate,
+    message,
+    mutateConfig,
+    mutateSkills,
+    onRefresh,
+    t
+  })
+  const skillsCliModal = useSkillsCliModalController({
+    message,
+    mutateSkills,
+    t
   })
 
   const skills = skillsRes?.skills ?? []
@@ -124,156 +107,20 @@ export function SkillsTab({
     ...registries.map(item => ({ label: item.id, value: item.id }))
   ], [registries, t])
 
-  const handleRefresh = async () => {
-    await Promise.all([mutateSkills(), marketSearch.mutate(), mutateConfig(), onRefresh()])
-  }
-
-  const handleInstall = async (item: SkillHubItem) => {
-    setInstallingId(item.id)
-    try {
-      await installSkillHubItem({
-        registry: item.registry,
-        plugin: item.installRef ?? item.name,
-        force: item.installed
-      })
-      await Promise.all([marketSearch.mutate(), mutateSkills()])
-      void message.success(t('knowledge.skills.installSuccess'))
-    } catch (error) {
-      void message.error(getApiErrorMessage(error, t('knowledge.skills.installFailed')))
-    } finally {
-      setInstallingId(null)
-    }
-  }
-
-  const handleImportArchive = async (file: File) => {
-    setImporting(true)
-    try {
-      const result = await importSkillArchive(file)
-      await Promise.all([mutateSkills(), onRefresh()])
-      void message.success(t('knowledge.skills.importSuccess', { count: result.fileCount }))
-    } catch (error) {
-      void message.error(getApiErrorMessage(error, t('knowledge.skills.importFailed')))
-    } finally {
-      setImporting(false)
-    }
-  }
-
-  const resolveSkillsCliRequest = React.useCallback(async () => {
-    try {
-      await skillsCliForm.validateFields(['source'])
-    } catch {
-      return undefined
-    }
-
-    const values = skillsCliForm.getFieldsValue()
-    const source = values.source.trim()
-    return {
-      source,
-      query: trimOptionalString(values.query),
-      registry: trimOptionalString(values.registry)
-    }
-  }, [skillsCliForm])
-
-  const runSkillsCliSearch = React.useCallback(async (nextLimit: number) => {
-    const request = await resolveSkillsCliRequest()
-    if (request == null) return false
-
-    const result = await searchSkillsCli({
-      limit: nextLimit,
-      source: request.source,
-      ...(request.query != null ? { query: request.query } : {}),
-      ...(request.registry != null ? { registry: request.registry } : {})
-    })
-
-    setSkillsCliHasSearched(true)
-    setSkillsCliLimit(nextLimit)
-    setSkillsCliItems(result.items)
-    setSkillsCliError(result.error ?? null)
-    setSkillsCliHasMore(result.hasMore === true && nextLimit < SKILLS_CLI_MAX_LIMIT)
-    setSkillsCliResetKey([request.source, request.query ?? '', request.registry ?? '', String(nextLimit)].join('\0'))
-    return true
-  }, [resolveSkillsCliRequest])
-
-  const handleSearchSkillsCli = async () => {
-    setSkillsCliSearching(true)
-    setSkillsCliLoadingMore(false)
-    try {
-      await runSkillsCliSearch(SKILLS_CLI_INITIAL_LIMIT)
-    } catch (error) {
-      void message.error(getApiErrorMessage(error, t('knowledge.skills.skillsCliSearchFailed')))
-    } finally {
-      setSkillsCliSearching(false)
-    }
-  }
-
-  const handleLoadMoreSkillsCli = async () => {
-    const nextLimit = Math.min(skillsCliLimit + SKILLS_CLI_LIMIT_STEP, SKILLS_CLI_MAX_LIMIT)
-    if (nextLimit === skillsCliLimit) return
-
-    setSkillsCliLoadingMore(true)
-    try {
-      await runSkillsCliSearch(nextLimit)
-    } catch (error) {
-      void message.error(getApiErrorMessage(error, t('knowledge.skills.skillsCliSearchFailed')))
-    } finally {
-      setSkillsCliLoadingMore(false)
-    }
-  }
-
-  const handleInstallSkillsCli = async (item: SkillHubItem) => {
-    const request = await resolveSkillsCliRequest()
-    if (request == null) return
-
-    setSkillsCliInstallingId(item.id)
-    try {
-      await installSkillsCliItem({
-        source: request.source,
-        skill: item.installRef ?? item.name,
-        force: item.installed,
-        ...(request.registry != null ? { registry: request.registry } : {})
-      })
-      await mutateSkills()
-      try {
-        await runSkillsCliSearch(skillsCliLimit)
-      } catch {
-        // Keep the install success state even if refreshing the source list fails afterwards.
-      }
-      void message.success(t('knowledge.skills.installSuccess'))
-    } catch (error) {
-      void message.error(getApiErrorMessage(error, t('knowledge.skills.installFailed')))
-    } finally {
-      setSkillsCliInstallingId(null)
-    }
-  }
-
-  const handleCloseSkillsCli = React.useCallback(() => {
-    setSkillsCliOpen(false)
-    setSkillsCliSearching(false)
-    setSkillsCliLoadingMore(false)
-    setSkillsCliInstallingId(null)
-    setSkillsCliItems([])
-    setSkillsCliError(null)
-    setSkillsCliHasMore(false)
-    setSkillsCliHasSearched(false)
-    setSkillsCliLimit(SKILLS_CLI_INITIAL_LIMIT)
-    setSkillsCliResetKey('')
-    skillsCliForm.resetFields()
-  }, [skillsCliForm])
-
   return (
     <TabContent className='knowledge-base-view__skills-tab'>
       <SkillsTabActions
-        importing={importing}
+        importing={actions.importing}
         leading={leading}
         viewMode={viewMode}
-        onRefresh={() => void handleRefresh()}
-        onImport={() => importInputRef.current?.click()}
+        onRefresh={() => void actions.handleRefresh()}
+        onImport={() => actions.importInputRef.current?.click()}
         onOpenConfig={() => navigate('/config?tab=plugins')}
         onViewModeChange={onViewModeChange}
       />
       <SkillArchiveInput
-        inputRef={importInputRef}
-        onSelect={(file) => void handleImportArchive(file)}
+        inputRef={actions.importInputRef}
+        onSelect={(file) => void actions.handleImportArchive(file)}
       />
       {viewMode === 'project' && (
         <ProjectSkillsList
@@ -286,7 +133,7 @@ export function SkillsTab({
       {viewMode === 'market' && (
         <SkillMarketView
           hubItems={marketFilters.filteredHubItems}
-          installingId={installingId}
+          installingId={actions.installingId}
           installFilter={installFilter}
           isLoading={marketSearch.isLoading && hubItems.length === 0}
           query={marketQuery}
@@ -299,8 +146,8 @@ export function SkillsTab({
           canLoadMore={marketSearch.canLoadMore}
           loadingMore={marketSearch.isValidating && hubItems.length > 0}
           onAddRegistry={() => registryModal.setOpen(true)}
-          onOpenSkillsCli={() => setSkillsCliOpen(true)}
-          onInstall={handleInstall}
+          onOpenSkillsCli={() => skillsCliModal.setOpen(true)}
+          onInstall={actions.handleInstall}
           onInstallFilterChange={onInstallFilterChange}
           onLoadMore={marketSearch.loadMore}
           onQueryChange={onMarketQueryChange}
@@ -318,20 +165,20 @@ export function SkillsTab({
         onClose={() => registryModal.setOpen(false)}
       />
       <SkillsCliModal
-        canLoadMore={skillsCliHasMore}
-        form={skillsCliForm}
-        hasSearched={skillsCliHasSearched}
-        installingId={skillsCliInstallingId}
-        items={skillsCliItems}
-        loadingMore={skillsCliLoadingMore}
-        open={skillsCliOpen}
-        resetKey={skillsCliResetKey}
-        searchError={skillsCliError}
-        searching={skillsCliSearching}
-        onClose={handleCloseSkillsCli}
-        onInstall={handleInstallSkillsCli}
-        onLoadMore={() => void handleLoadMoreSkillsCli()}
-        onSearch={() => void handleSearchSkillsCli()}
+        canLoadMore={skillsCliModal.hasMore}
+        form={skillsCliModal.form}
+        hasSearched={skillsCliModal.hasSearched}
+        installingId={skillsCliModal.installingId}
+        items={skillsCliModal.items}
+        loadingMore={skillsCliModal.loadingMore}
+        open={skillsCliModal.open}
+        resetKey={skillsCliModal.resetKey}
+        searchError={skillsCliModal.searchError}
+        searching={skillsCliModal.searching}
+        onClose={skillsCliModal.handleClose}
+        onInstall={skillsCliModal.handleInstall}
+        onLoadMore={() => void skillsCliModal.handleLoadMore()}
+        onSearch={() => void skillsCliModal.handleSearch()}
       />
     </TabContent>
   )
