+++ eslint.max-lines-baseline.mjs
@@ -8,7 +8,6 @@ export default [
   '.ai/rfcs/0001-gemini-cli-adapter.md',
   '.github/workflows/desktop-package.yml',
   'apps/cli/src/commands/list.ts',
-  'apps/cli/src/commands/skills.ts',
   'apps/cli/src/commands/run/command.ts',
   'apps/cli/src/session-cache.ts',
   'apps/client/src/api/sessions.ts',
@@ -35,7 +34,6 @@ export default [
   'apps/client/src/components/config/record-editors/ModelServicesRecordEditor.tsx',
   'apps/client/src/components/ConfigView.tsx',
   'apps/client/src/components/knowledge-base/KnowledgeBaseView.tsx',
-  'apps/client/src/components/knowledge-base/components/SkillsTab.tsx',
   'apps/client/src/components/Sidebar.tsx',
   'apps/client/src/components/server-connection/ServerConnectionGate.tsx',
   'apps/client/src/components/sidebar/SessionItem.tsx',
@@ -109,10 +107,7 @@ export default [
   'packages/utils/src/create-logger.ts',
   'packages/utils/src/model-selection.ts',
   'packages/utils/src/plugin-resolver.ts',
-  'packages/utils/src/project-skills.ts',
-  'packages/utils/src/skills-cli.ts',
   'packages/workspace-assets/src/bundle-internal.ts',
-  'packages/workspace-assets/src/skills-cli-dependency.ts',
   'packages/workspace-assets/__tests__/prompt-selection.spec.ts',
   'packages/workspace-assets/src/selection-internal.ts',
   'packages/channels/lark/__tests__/lark-channel.test.ts',
